modified adding of datamine column
