+++ server/src/parse-datapacks.ts
@@ -62,8 +62,7 @@ import {
   defaultSequenceSettings,
   assertSequence,
   SequenceSettings,
-  createDefaultColumnHeaderProps,
-  getValidFontOptions
+  allFontOptions
 } from "@tsconline/shared";
 import {
   grabFilepaths,
@@ -1881,4 +1880,49 @@ function handlePointFields(point: Point, loneColumns: Map<string, ColumnInfo>, u
     point.name,
     createLoneColumn(headerInfo, getValidFontOptions("Point"), units, subPointInfo, "Point", columnSpecificSettings)
   );
+}
+
+export function createDefaultColumnHeaderProps(overrides: Partial<ColumnHeaderProps> = {}): ColumnHeaderProps {
+  const defaultRGB: RGB = { r: 255, g: 255, b: 255 };
+  const defaultProps: ColumnHeaderProps = {
+    name: "",
+    minAge: Number.MAX_VALUE,
+    maxAge: Number.MIN_VALUE,
+    enableTitle: true,
+    on: true,
+    width: 100,
+    popup: "",
+    rgb: defaultRGB
+  };
+
+  return { ...defaultProps, ...overrides };
+}
+
+export function getValidFontOptions(type: DisplayedColumnTypes): ValidFontOptions[] {
+  switch (type) {
+    case "Block":
+    case "Zone":
+      return ["Column Header", "Age Label", "Zone Column Label"];
+    case "Chron":
+      return ["Column Header", "Age Label"];
+    case "Event":
+      return ["Column Header", "Age Label", "Event Column Label", "Uncertainty Label", "Range Label"];
+    case "Facies":
+      return ["Column Header", "Age Label", "Uncertainty Label"];
+    case "Point":
+      return ["Column Header", "Point Column Scale Label"];
+    case "Range":
+      return [...allFontOptions];
+    case "Sequence":
+      return ["Column Header", "Age Label", "Sequence Column Label"];
+    case "Ruler":
+    case "AgeAge":
+      return ["Column Header", "Ruler Label"];
+    case "Transect":
+      return ["Column Header"];
+    case "Freehand":
+      return ["Column Header"];
+    default:
+      return ["Column Header"];
+  }
 }
